Refactor: Extract hooks, utils, and improve code organization
