+++ SMS_V.2.0/src/components/features/ExchangeRateExample.tsx
@@ -1,7 +1,7 @@
 import React, { useState } from 'react';
 import { CurrencyDollarIcon, ArrowDownIcon } from '@heroicons/react/24/outline';
 import { cn } from '../../lib/utils';
-import { useExchangeRateContext } from '../../contexts/ExchangeRateContext';
+import { useExchangeRateContext } from '../../hooks/useExchangeRateContext';
 
 interface ExchangeRateExampleProps {
   className?: string;
