+++ SMS_V.2.0/src/contexts/AuthContext.ts
@@ -1,0 +1,14 @@
+import { createContext } from 'react';
+import type { User, Session, AuthError } from '@supabase/supabase-js';
+
+interface AuthContextType {
+  user: User | null;
+  session: Session | null;
+  loading: boolean;
+  signIn: (email: string, password: string) => Promise<{ error: AuthError | null }>;
+  signUp: (email: string, password: string) => Promise<{ error: AuthError | null }>;
+  signOut: () => Promise<void>;
+  signInWithGoogle: () => Promise<{ error: AuthError | null }>;
+}
+
+export const AuthContext = createContext<AuthContextType | undefined>(undefined);
