+++ SMS_V.2.0/src/hooks/useExchangeRateContext.ts
@@ -1,0 +1,10 @@
+import { useContext } from 'react';
+import { ExchangeRateContext } from '../contexts/ExchangeRateContext';
+
+export const useExchangeRateContext = () => {
+  const context = useContext(ExchangeRateContext);
+  if (context === undefined) {
+    throw new Error('useExchangeRateContext must be used within an ExchangeRateProvider');
+  }
+  return context;
+};
